# [11_35] Performance tuning on font_database_styles