--- a/src/Graphics/Fonts/font_database.cpp
+++ b/src/Graphics/Fonts/font_database.cpp
@@ -123,9 +123,11 @@ font_database_build_styles (tree t) {
     string style = t[1]->label;
     if (font_cache_family_to_styles->contains (family)) {
       array<string> styles= font_cache_family_to_styles[family];
-      styles << style;
-      merge_sort_leq<string, locase_less_eq_operator> (styles);
-      font_cache_family_to_styles (family)= styles;
+      if (!contains (style, styles)) {
+        styles << style;
+        merge_sort_leq<string, locase_less_eq_operator> (styles);
+        font_cache_family_to_styles (family)= styles;
+      }
     }
     else {
       array<string> styles= array<string> ();