## What
On some big doc:
6467 ms -> 5223 ms
3465 ms -> 3027 ms

## Why
font_database_styles is computing every time, very slow